Implement add book.
- Added controller, service and repository for each of Currency, Book and BookInstance.
- Code restructuring.

+++ src/main/java/com/jk/ptk/f/user/UserServiceHelper.java
@@ -52,7 +52,7 @@ class UserServiceHelper {
 	 *           the specified id
 	 * @return a country as identified by the specified {@code id}
 	 */
-	Country findCountry(Long id) {
+	Country findCountry(Integer id) {
 		return countryRepository.find(id);
 	}
